# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "DevEventTrace": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "payload": {
            "title": "Payload"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "sequence",
          "event_type",
          "payload",
          "created_at"
        ],
        "title": "DevEventTrace",
        "type": "object"
      },
      "DevRunDetail": {
        "additionalProperties": false,
        "properties": {
          "config_snapshot": {
            "$ref": "#/components/schemas/DevSnapshot"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/DevEventTrace"
            },
            "title": "Events",
            "type": "array"
          },
          "input_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DevSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_text": {
            "title": "Request Text",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "run": {
            "$ref": "#/components/schemas/DevRunSummary"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/DevStepTrace"
            },
            "title": "Steps",
            "type": "array"
          },
          "terminal_invariant": {
            "$ref": "#/components/schemas/TerminalInvariant"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/DevToolTrace"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "run",
          "request_text",
          "input_snapshot",
          "config_snapshot",
          "result",
          "steps",
          "tools",
          "events",
          "terminal_invariant"
        ],
        "title": "DevRunDetail",
        "type": "object"
      },
      "DevRunListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DevRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "DevRunListResponse",
        "type": "object"
      },
      "DevRunSummary": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "graph_version": {
            "title": "Graph Version",
            "type": "string"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "replay_of_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replay Of Run Id"
          },
          "resolved_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Intent"
          },
          "result_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Kind"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_cost_cny": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Cost Cny",
            "type": "string"
          },
          "total_latency_ms": {
            "title": "Total Latency Ms",
            "type": "integer"
          },
          "total_tokens_in": {
            "title": "Total Tokens In",
            "type": "integer"
          },
          "total_tokens_out": {
            "title": "Total Tokens Out",
            "type": "integer"
          },
          "user_ref": {
            "title": "User Ref",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "replay_of_run_id",
          "user_ref",
          "status",
          "result_kind",
          "resolved_intent",
          "graph_version",
          "model_id",
          "total_tokens_in",
          "total_tokens_out",
          "total_cost_cny",
          "total_latency_ms",
          "fallback_reason",
          "error_code",
          "created_at",
          "finished_at"
        ],
        "title": "DevRunSummary",
        "type": "object"
      },
      "DevSnapshot": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "title": "Data"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "data",
          "sha256"
        ],
        "title": "DevSnapshot",
        "type": "object"
      },
      "DevStepTrace": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "input_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Hash"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "node_name": {
            "title": "Node Name",
            "type": "string"
          },
          "output_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Hash"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tokens_in": {
            "title": "Tokens In",
            "type": "integer"
          },
          "tokens_out": {
            "title": "Tokens Out",
            "type": "integer"
          },
          "trace_data": {
            "title": "Trace Data"
          }
        },
        "required": [
          "sequence",
          "node_name",
          "attempt",
          "status",
          "prompt_version",
          "model_id",
          "tokens_in",
          "tokens_out",
          "latency_ms",
          "input_hash",
          "output_hash",
          "trace_data",
          "error_code"
        ],
        "title": "DevStepTrace",
        "type": "object"
      },
      "DevToolTrace": {
        "additionalProperties": false,
        "properties": {
          "args": {
            "title": "Args"
          },
          "args_hash": {
            "title": "Args Hash",
            "type": "string"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "result_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Hash"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "step_id": {
            "format": "uuid",
            "title": "Step Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "tool_call_id": {
            "format": "uuid",
            "title": "Tool Call Id",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "tool_call_id",
          "step_id",
          "tool_name",
          "contract_version",
          "round",
          "args",
          "args_hash",
          "result",
          "result_hash",
          "provider",
          "latency_ms",
          "success",
          "error_code"
        ],
        "title": "DevToolTrace",
        "type": "object"
      },
...
      "EvalDatasetListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvalDatasetSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "EvalDatasetListResponse",
        "type": "object"
      },
      "EvalDatasetSummary": {
        "additionalProperties": false,
        "properties": {
          "case_count": {
            "minimum": 1.0,
            "title": "Case Count",
            "type": "integer"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "required": [
          "dataset_id",
          "case_count",
          "description"
        ],
        "title": "EvalDatasetSummary",
        "type": "object"
      },
      "EvalExperimentResponse": {
        "additionalProperties": false,
        "properties": {
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "report": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Report"
          },
          "status": {
            "enum": [
              "completed",
              "not_found"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "status",
          "report"
        ],
        "title": "EvalExperimentResponse",
        "type": "object"
      },
      "EvalStartRequest": {
        "additionalProperties": false,
        "properties": {
          "case_limit": {
            "anyOf": [
              {
                "maximum": 30.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Limit"
          },
          "dataset_id": {
            "default": "stage5-v1",
            "title": "Dataset Id",
            "type": "string"
          }
        },
        "title": "EvalStartRequest",
        "type": "object"
      },
      "EvalStartResponse": {
        "additionalProperties": false,
        "properties": {
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "status": {
            "const": "completed",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "status"
        ],
        "title": "EvalStartResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplayRequest": {
        "additionalProperties": false,
        "properties": {
          "tool_mode": {
            "default": "fixture",
            "enum": [
              "fixture",
              "live"
            ],
            "title": "Tool Mode",
            "type": "string"
          }
        },
        "title": "ReplayRequest",
        "type": "object"
      },
      "ReplayResponse": {
        "additionalProperties": false,
        "properties": {
          "deterministic": {
            "title": "Deterministic",
            "type": "boolean"
          },
          "replay_of_run_id": {
            "format": "uuid",
            "title": "Replay Of Run Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "replay_of_run_id",
          "status",
          "deterministic"
        ],
        "title": "ReplayResponse",
        "type": "object"
      },
...
      "TerminalInvariant": {
        "additionalProperties": false,
        "properties": {
          "terminal_count": {
            "title": "Terminal Count",
            "type": "integer"
          },
          "terminal_is_last": {
            "title": "Terminal Is Last",
            "type": "boolean"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "terminal_count",
          "terminal_is_last",
          "valid"
        ],
        "title": "TerminalInvariant",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Career Planning Buddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dev/evals/datasets": {
      "get": {
        "operationId": "list_eval_datasets_api_v1_dev_evals_datasets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalDatasetListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Eval Datasets",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/dev/evals/experiments": {
      "post": {
        "operationId": "start_eval_api_v1_dev_evals_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Eval",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/dev/evals/experiments/{experiment_id}": {
      "get": {
        "operationId": "get_eval_api_v1_dev_evals_experiments__experiment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalExperimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Eval",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/dev/runs": {
      "get": {
        "operationId": "list_runs_api_v1_dev_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "result_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Kind"
            }
          },
          {
            "in": "query",
            "name": "error_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/dev/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_dev_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/dev/runs/{run_id}/replay": {
      "post": {
        "operationId": "replay_run_api_v1_dev_runs__run_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replay Run",
        "tags": [
          "developer"
        ]
      }
    },
...
  }
}